A series of fixes related to scheduling of group meetings
