+++ tools/lib/project.mjs
@@ -945,3 +945,41 @@ export function validateProject(project) {
 
   return errors;
 }
+
+
+/**
+ * Convert the project to a simplified JSON data structure
+ * (suitable for tests but also for debugging)
+ */
+export function convertProjectToJSON(project) {
+  const toNameList = list => list.map(item => item.name);
+  const data = {
+    title: project.title,
+    description: project.description
+  };
+  if (project.allowMultipleMeetings) {
+    data.allowMultipleMeetings = true;
+  }
+  for (const list of ['days', 'rooms', 'slots', 'labels']) {
+    data[list] = toNameList(project[list]);
+  }
+
+  data.sessions = project.sessions.map(session => {
+    const simplified = {
+      number: session.number,
+      title: session.title,
+      author: session.author.login,
+      body: session.body,
+    };
+    if (session.labels.length !== 1 || session.labels[0] !== 'session') {
+      simplified.labels = session.labels;
+    }
+    for (const field of ['day', 'room', 'slot', 'meeting']) {
+      if (session[field]) {
+        simplified[field] = session[field];
+      }
+    }
+    return simplified;
+  });
+  return data;
+}
